Add some test coverage
 * a unit test for the HTTP headers to send

 * an integration test that hits the live page
and adds a real YellowPage example from the
documentation, with updates to the CSS selectors
based on markup changes since the docs were written.

Instead of a live page, mocha could be used to 
serve a stub HTML response, but that wasn't done
here.

After bundling, run "rake test" to run the tests.